Add support for https (always redirect http -> https)
Sources that contain relevant documentation have been added along with
the rest of the links. To use https, we configure nginx to listen to
port 443 and provide the paths of our certificates (both private and
public). We add a new server block, in which we redirect all http
connections to https. Lastly, we now also include a local domain name.