Fix incorrect tag name bug to properly render CSS
Through testing it was fount that the CSS was not properly rendering and
that the 'player' website was not responsive for mobile devices. This
was because in the player.html file there was mistake where when linking
the CSS file, the tag was changed from 'style' to 'link'. Also mores
fields that make the website more responsive to the 'video' class in the
CSS file.